Changing the old code into functions
I still have to create functions for some old codes and analyse them further. I would also like to automate the fact of not always having to write the qID but to have a variable that chooses all the same queries and apply the functions for different queries, this way I would have an overview of the whole run file.